refactor: 피드백적용 - application 패키지명 service 로 변경
