+++ src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.ui;
 
-import kitchenpos.menu.application.MenuGroupService;
+import kitchenpos.menu.service.MenuGroupService;
 import kitchenpos.menu.dto.MenuGroupRequest;
 import kitchenpos.menu.dto.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
