+++ src/main/java/kitchenpos/order/service/TableGroupService.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.application;
+package kitchenpos.order.service;
 
 import kitchenpos.order.domain.entity.*;
 import kitchenpos.order.domain.value.OrderStatus;
