+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.ui;
 
-import kitchenpos.order.application.OrderService;
+import kitchenpos.order.service.OrderService;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
 import org.springframework.http.ResponseEntity;
