+++ src/main/java/kitchenpos/order/ui/TableGroupRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.ui;
 
-import kitchenpos.order.application.TableGroupService;
+import kitchenpos.order.service.TableGroupService;
 import kitchenpos.order.dto.TableGroupRequest;
 import kitchenpos.order.dto.TableGroupResponse;
 import org.springframework.http.ResponseEntity;
