+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.product.ui;
 
-import kitchenpos.product.application.ProductService;
+import kitchenpos.product.service.ProductService;
 import kitchenpos.product.dto.ProductRequest;
 import kitchenpos.product.dto.ProductResponse;
 import org.springframework.http.ResponseEntity;
